Move classes/files to codec package
