+++ src/main/java/com/fauna/codec/MapOf.java
@@ -1,4 +1,4 @@
-package com.fauna.serialization.generic;
+package com.fauna.codec;
 
 import java.lang.reflect.Type;
 import java.util.Map;
