+++ src/main/java/com/fauna/codec/Parameterized.java
@@ -1,4 +1,4 @@
-package com.fauna.serialization.generic;
+package com.fauna.codec;
 
 /**
  * A helper class for static access to parameterized generics for deserialization.
