+++ src/main/java/com/fauna/codec/ParameterizedOf.java
@@ -1,4 +1,4 @@
-package com.fauna.serialization.generic;
+package com.fauna.codec;
 
 import java.lang.reflect.ParameterizedType;
