+++ src/main/java/com/fauna/codec/UTF8FaunaGenerator.java
@@ -1,4 +1,4 @@
-package com.fauna.serialization;
+package com.fauna.codec;
 
 
 import com.fasterxml.jackson.core.JsonFactory;
