+++ src/main/java/com/fauna/codec/UTF8FaunaParser.java
@@ -1,4 +1,4 @@
-package com.fauna.serialization;
+package com.fauna.codec;
 
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonParser;
