+++ src/main/java/com/fauna/codec/codecs/ClassCodec.java
@@ -8,8 +8,8 @@ import com.fauna.codec.CodecProvider;
 import com.fauna.enums.FaunaTokenType;
 import com.fauna.exception.ClientException;
 import com.fauna.mapping.FieldInfo;
-import com.fauna.serialization.UTF8FaunaGenerator;
-import com.fauna.serialization.UTF8FaunaParser;
+import com.fauna.codec.UTF8FaunaGenerator;
+import com.fauna.codec.UTF8FaunaParser;
 
 import java.io.IOException;
 import java.lang.reflect.Constructor;
