+++ src/main/java/com/fauna/codec/codecs/DynamicCodec.java
@@ -2,8 +2,9 @@ package com.fauna.codec.codecs;
 
 import com.fauna.codec.Codec;
 import com.fauna.codec.CodecProvider;
+import com.fauna.codec.UTF8FaunaGenerator;
+import com.fauna.codec.UTF8FaunaParser;
 import com.fauna.exception.ClientException;
-import com.fauna.serialization.*;
 import com.fauna.types.*;
 
 import java.io.IOException;
