+++ src/main/java/com/fauna/codec/codecs/InstantCodec.java
@@ -1,8 +1,8 @@
 package com.fauna.codec.codecs;
 
 import com.fauna.exception.ClientException;
-import com.fauna.serialization.UTF8FaunaGenerator;
-import com.fauna.serialization.UTF8FaunaParser;
+import com.fauna.codec.UTF8FaunaGenerator;
+import com.fauna.codec.UTF8FaunaParser;
 
 import java.io.IOException;
 import java.time.Instant;
