+++ src/main/java/com/fauna/codec/codecs/NullableCodec.java
@@ -4,8 +4,8 @@ import com.fauna.codec.Codec;
 import com.fauna.enums.FaunaTokenType;
 import com.fauna.exception.ClientException;
 import com.fauna.exception.NullDocumentException;
-import com.fauna.serialization.UTF8FaunaGenerator;
-import com.fauna.serialization.UTF8FaunaParser;
+import com.fauna.codec.UTF8FaunaGenerator;
+import com.fauna.codec.UTF8FaunaParser;
 import com.fauna.types.NonNull;
 import com.fauna.types.NullDoc;
 import com.fauna.types.Nullable;
