+++ src/main/java/com/fauna/codec/codecs/PageCodec.java
@@ -3,8 +3,8 @@ package com.fauna.codec.codecs;
 import com.fauna.codec.Codec;
 import com.fauna.enums.FaunaTokenType;
 import com.fauna.exception.ClientException;
-import com.fauna.serialization.UTF8FaunaGenerator;
-import com.fauna.serialization.UTF8FaunaParser;
+import com.fauna.codec.UTF8FaunaGenerator;
+import com.fauna.codec.UTF8FaunaParser;
 import com.fauna.types.Page;
 
 import java.io.IOException;
