+++ src/main/java/com/fauna/interfaces/IDeserializer.java
@@ -1,7 +1,7 @@
 package com.fauna.interfaces;
 
 
-import com.fauna.serialization.UTF8FaunaParser;
+import com.fauna.codec.UTF8FaunaParser;
 import java.io.IOException;
 
 public interface IDeserializer<T> {
