+++ src/main/java/com/fauna/serialization/CheckedDeserializer.java
@@ -1,5 +1,6 @@
 package com.fauna.serialization;
 
+import com.fauna.codec.UTF8FaunaParser;
 import com.fauna.exception.ClientException;
 
 import java.io.IOException;
