+++ src/main/java/com/fauna/serialization/OptionalDeserializer.java
@@ -1,7 +1,7 @@
 package com.fauna.serialization;
 
+import com.fauna.codec.UTF8FaunaParser;
 import com.fauna.enums.FaunaTokenType;
-import com.fauna.exception.ClientException;
 import com.fauna.interfaces.IDeserializer;
 
 import java.io.IOException;
