+++ src/test/java/com/fauna/client/PageIteratorTest.java
@@ -8,8 +8,8 @@ import com.fauna.response.QueryFailure;
 import com.fauna.response.QuerySuccess;
 import com.fauna.serialization.Deserializer;
 import com.fauna.serialization.PageDeserializer;
-import com.fauna.serialization.generic.PageOf;
-import com.fauna.serialization.generic.ParameterizedOf;
+import com.fauna.codec.PageOf;
+import com.fauna.codec.ParameterizedOf;
 import com.fauna.types.Page;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
