+++ src/test/java/com/fauna/codec/UTF8FaunaParserTest.java
@@ -1,9 +1,10 @@
-package com.fauna.serialization;
+package com.fauna.codec;
 
 import static org.junit.Assert.*;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import com.fauna.codec.UTF8FaunaParser;
 import com.fauna.enums.FaunaTokenType;
 import com.fauna.types.Module;
 import com.fauna.exception.ClientException;
